Move password hashing helper function into util
